fix(sieve): fix int64 overflow in UV computation for large primes
f0_mod * m64 (or r64) can overflow int64_t when p ~ 2^32 since
(2^32-1)^2 > INT64_MAX. Uses __int128_t for intermediate products
in both compute_rational_uv and compute_algebraic_uv.

+++ include/gnfs/sieve/lattice_sieve.hpp
@@ -278,9 +278,9 @@ private:
         int64_t e1_mod = mod_reduce(basis.e1);
         int64_t f1_mod = mod_reduce(basis.f1);
         int64_t m64 = static_cast<int64_t>(m_mod_p);
-        // f_mod * m64 fits in int64: both < p < 2^32
-        int64_t u = (e0_mod - f0_mod * m64 % p64 + p64) % p64;
-        int64_t v = (e1_mod - f1_mod * m64 % p64 + p64) % p64;
+        // Use __int128_t for intermediate product to avoid overflow when p ~ 2^32
+        int64_t u = (e0_mod - static_cast<int64_t>(static_cast<__int128_t>(f0_mod) * m64 % p64) + p64) % p64;
+        int64_t v = (e1_mod - static_cast<int64_t>(static_cast<__int128_t>(f1_mod) * m64 % p64) + p64) % p64;
         return {u, v};
     }
 
@@ -299,8 +299,9 @@ private:
         int64_t e1_mod = mod_reduce(basis.e1);
         int64_t f1_mod = mod_reduce(basis.f1);
         int64_t r64 = static_cast<int64_t>(r);
-        int64_t u = (e0_mod - f0_mod * r64 % p64 + p64) % p64;
-        int64_t v = (e1_mod - f1_mod * r64 % p64 + p64) % p64;
+        // Use __int128_t for intermediate product to avoid overflow when p ~ 2^32
+        int64_t u = (e0_mod - static_cast<int64_t>(static_cast<__int128_t>(f0_mod) * r64 % p64) + p64) % p64;
+        int64_t v = (e1_mod - static_cast<int64_t>(static_cast<__int128_t>(f1_mod) * r64 % p64) + p64) % p64;
         return {u, v};
     }
